requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.5.0
assertpy==1.1
astor==0.8.1
attrs==23.1.0
black==23.7.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
flake8==6.1.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pyflakes==3.1.0
pyparsing==3.1.0
pytest==7.4.0
pytest-ast-transformer==1.0.3
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
texttable==1.6.7
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
pydantic==2.5.2
faker
